+++ .github/settings.yml
@@ -9,19 +9,36 @@ repository:
   allow_merge_commit: false
   allow_rebase_merge: false
 labels:
+  - name: "breaking"
+    color: ffcc00
+    description: "Breaking change."
   - name: "bug"
-    color: "d73a4a"
+    color: d73a4a
+    description: "Something isn't working."
+  - name: "dependencies"
+    color: 0366d6
+    description: "Pull requests that update a dependency file."
   - name: "documentation"
-    color: "0075ca"
+    color: 0075ca
+    description: "Improvements or additions to documentation."
   - name: "duplicate"
-    color: "cfd3d7"
+    color: cfd3d7
+    description: "This issue or pull request already exists."
   - name: "enhancement"
-    color: "a2eeef"
+    color: a2eeef
+    description: "New feature or request."
   - name: "invalid"
-    color: "e4e669"
+    color: e4e669
+    description: "This doesn't seem right."
   - name: "stale"
-    color: "930191"
+    color: 930191
+    description: "Stale issues and PRs."
   - name: "question"
-    color: "dd9c2c"
+    color: dd9c2c
+    description: "Further information is requested."
+  - name: "tests"
+    color: bfd4f2
+    description: "CI, CD and testing related changes."
   - name: "wontfix"
-    color: "ffffff"
+    color: ffffff
+    description: "This will not be worked on."
