+++ .github/workflows/pre-commit-update.yml
@@ -1,0 +1,29 @@
+name: Pre-commit auto-update
+on:
+  schedule:
+    - cron: '0 0 * * *'
+jobs:
+  auto-update:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - name: "Set up Python"
+        uses: actions/setup-python@v2
+        with:
+          python-version: 3.8
+      - name: "Install pre-commit"
+        run: pip install pre-commit
+      - name: "Run pre-commit autoupdate"
+        run: pre-commit autoupdate
+      - name: "Create Pull Request"
+        uses: peter-evans/create-pull-request@v2
+        with:
+          token: ${{ secrets.CPR_GITHUB_TOKEN }}
+          branch: update/pre-commit-autoupdate
+          title: "Auto-update pre-commit hooks"
+          commit-message: "Auto-update pre-commit hooks"
+          body: |
+            Update versions of tools in pre-commit
+            configs to latest version
+          labels: dependencies
+u
