+++ .pre-commit-config.yaml
@@ -1,17 +1,17 @@
 repos:
   - repo: https://github.com/asottile/pyupgrade
-    rev: v2.4.1
+    rev: v2.7.4
     hooks:
       - id: pyupgrade
         args: [--py37-plus]
   - repo: https://github.com/psf/black
-    rev: 19.10b0
+    rev: 20.8b1
     hooks:
       - id: black
         args:
           - --safe
           - --quiet
-        files: ^custom_components/.+\.py$
+        files: ^(custom_components|bin|tests)/.+\.py$
   - repo: https://gitlab.com/pycqa/flake8
     rev: 3.8.1
     hooks:
@@ -19,7 +19,20 @@ repos:
         additional_dependencies:
           - flake8-docstrings==1.5.0
           - pydocstyle==5.0.2
-        files: ^custom_components/.+\.py$
+        files: ^(custom_components|bin|tests)/.+\.py$
+  - repo: https://github.com/PyCQA/bandit
+    rev: 1.7.0
+    hooks:
+      - id: bandit
+        args:
+          - --quiet
+          - --format=custom
+          - --configfile=tests/bandit.yaml
+        files: ^(custom_components|bin|tests)/.+\.py$
+  - repo: https://github.com/pre-commit/mirrors-isort
+    rev: v5.6.4
+    hooks:
+      - id: isort
   - repo: https://github.com/pre-commit/pre-commit-hooks
     rev: v2.5.0
     hooks:
@@ -27,9 +40,9 @@ repos:
         stages: [manual]
       - id: check-json
       - id: trailing-whitespace
-#      - id: no-commit-to-branch
-#        args:
-#          - --branch=master
+      - id: no-commit-to-branch
+        args:
+          - --branch=master
   - repo: local
     hooks:
       # Run mypy through our wrapper script in order to get the possible
