+++ bin/gen_releasenotes
@@ -1,14 +1,14 @@
 #!/usr/bin/env python3
 """Helper script to generate release notes."""
 import argparse
+from datetime import datetime
 import logging
 import os
 import re
 import subprocess
-from datetime import datetime
 from typing import List
 
-from github import Github, Repository, Tag, GithubException
+from github import Github, GithubException, Repository, Tag
 from packaging.version import Version
 
 # http://docs.python.org/2/howto/logging.html#library-config
@@ -19,7 +19,7 @@ logging.basicConfig(level=logging.CRITICAL)
 
 _LOGGER = logging.getLogger(__name__)
 
-VERSION = "1.1.2"
+VERSION = "1.1.4"
 
 ROOT = os.path.dirname(os.path.abspath(f"{__file__}/.."))
 
@@ -43,7 +43,7 @@ def get_commits(repo: Repository, since: datetime, until: datetime):
     try:
         dev = repo.get_branch("develop")
         dev_commits = repo.get_commits(sha=dev.commit.sha, since=since, until=until)
-        if len(dev_commits) > len(commits):
+        if dev_commits.totalCount > commits.totalCount:
             commits = dev_commits
     except GithubException:
         pass
@@ -55,7 +55,7 @@ def get_commits(repo: Repository, since: datetime, until: datetime):
 def get_release_tags(repo: Repository) -> List[Tag.Tag]:
     """Get list of all release tags from repository."""
     reg = re.compile(
-        r"^v?[0-9]+\.[0-9]+\.[0-9]+"
+        r"^v?[0-9]+\.[0-9]+(?:\.[0-9]+)?"
         r"(?:-[0-9A-Za-z-]+(?:\.[0-9A-Za-z-]+)*)?(?:\+[0-9A-Za-z-]+)?$"
     )
     tags = list(filter(lambda tag: re.match(reg, tag.name), list(repo.get_tags())))
