+++ requirements-dev.txt
@@ -1,11 +1,11 @@
--r requirements.txt
+-r requirements-test.txt
 black==20.08b1
-flake8~=3.8
-mypy~=0.800
-packaging~=20.9
-pre-commit~=2.10
-PyGithub~=1.54
-pylint~=2.6
+flake8==3.8.4
+mypy==0.790
+packaging==20.9
+pre-commit==2.8.2
+PyGithub==1.54.1
+pylint==2.6.0
 pylint-strict-informational==0.1
-pyupgrade~=2.10
-yamllint~=1.26
+pyupgrade==2.10.0
+yamllint==1.26.0
