+++ requirements-test.txt
@@ -1,0 +1,4 @@
+-r requirements.txt
+pytest==6.2.2
+pytest-cov==2.10.1
+pytest-homeassistant-custom-component==0.2.0
