+++ requirements.txt
@@ -1,4 +1,6 @@
-geopy~=2.1
-geohash2~=1.1
-voluptuous~=0.12
-homeassistant~=2021.2
+aiohttp==3.7.3
+defusedxml==0.6.0
+geopy==2.1.0
+geohash2==1.1
+voluptuous==0.12.1
+homeassistant==2021.2.3
