+++ setup.cfg
@@ -1,14 +1,67 @@
+[coverage:run]
+source =
+  custom_components
+
+[coverage:report]
+exclude_lines =
+    pragma: no cover
+    raise NotImplemented()
+    if __name__ == '__main__':
+    main()
+fail_under = 93
+show_missing = true
+
+[tool:pytest]
+testpaths = tests
+norecursedirs =
+    .git
+    dev-config
+addopts =
+    --strict-markers
+    --cov=custom_components
+
 [flake8]
+# https://github.com/ambv/black#line-length
 max-line-length = 88
+# E501: line too long
+# W503: Line break occurred before a binary operator
+# E203: Whitespace before ':'
+# D202 No blank lines allowed after function docstring
+# W504 line break after binary operator
+ignore =
+    E501,
+    W503,
+    E203,
+    D202,
+    W504
 exclude =
     .git,
     __pycache__,
-    dev_config
+    dev-config
+
+[isort]
+# https://github.com/timothycrosley/isort
+# https://github.com/timothycrosley/isort/wiki/isort-Settings
+# splits long import on multiple lines indented by 4 spaces
+multi_line_output = 3
+include_trailing_comma=True
+force_grid_wrap=0
+use_parentheses=True
+line_length=88
+indent = "    "
+# will group `import x` and `from x import` of the same module.
+force_sort_within_sections = true
+sections = FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER
+default_section = THIRDPARTY
+known_first_party = custom_components,tests
+forced_separate = tests
+combine_as_imports = true
 
 [mypy]
-ignore_missing_imports = True
-
-[semantic_release]
-version_variable=custom_components/gismeteo/__init__.py:VERSION
-upload_to_pypi=false
-upload_to_release=false
+python_version = 3.7
+ignore_errors = true
+follow_imports = silent
+ignore_missing_imports = true
+warn_incomplete_stub = true
+warn_redundant_casts = true
+warn_unused_configs = true
