+++ tests/fixtures/city_location.xml
@@ -1,0 +1,1 @@
+<?xml version="1.0" encoding="utf-8"?><document><item id="167413" lat="55.5914" lng="37.743698" tzone="180" n="Razvilka" country_id="156" distance="1.68" country_name="Russia" district_id="251" district_name="Moscow Oblast"  kind="T"/></document>
