+++ tests/fixtures/forecast_data.xml
@@ -1,0 +1,48 @@
+<?xml version="1.0" encoding="utf-8"?><weather><location id="167413" name="Razvilka" name_r="in Razvilka" country_id="156"
+              country_name="Russia" district_id="251" district_name="Moscow Oblast" kind="T" lat="55.59" lng="37.74"  tzone="180" cur_time="2021-02-21T16:16:00" nowcast_url="/nowcast-razvilka-167413/" nowcast_teaser="true"><fact valid="2021-02-21T16:00:00" tod="2" risem="459" setm="1067" durm="608" sunrise="1613893140" sunset="1613929620"><values t="-7" tflt="-7" tcflt="-11.5"  p="746" ws="3" wd="5" hum="86" hi="-12" cl="3" pt="2" pr="1" prflt="0.3" ts="0" water_t="3" icon="c4.s1" descr="Mainly cloudy, light snow" grade="3" ph="71"  /></fact><day date="2021-02-20" risem="461" setm="1064" durm="603" sunrise="1613806860" sunset="1613843040"></day><day date="2021-02-21" risem="459" setm="1067" durm="608" sunrise="1613893140" sunset="1613929620" tmin="-10" tmax="-7" pmin="746"
+                 pmax="751" wsmin="1" wsmax="3" hummin="79" hummax="89" cl="3" pt="2" pr="2" ts="0" icon="c4.s2" descr="Mainly cloudy, heavy snow" p="746"
+                 ws="3" wd="5" hum="84" grademax="3" prflt="5.8"  gust_speed="5"><forecast valid="2021-02-21T00:00:00" tod="-1"><values t="-10" p="751" ws="2" wd="3" hum="85" hi="-14" cl="3" pt="2" pr="1" prflt="0.2" ts="0"
+                                icon="c4.s1" descr="Mainly cloudy, light snow"
+                                grade="3"  /></forecast><forecast valid="2021-02-21T03:00:00" tod="0"><values t="-10" p="750" ws="1" wd="3" hum="79" hi="-10" cl="2" pt="0" pr="0" prflt="0" ts="0"
+                                icon="n.c3" descr="Cloudy"
+                                grade="4"  /></forecast><forecast valid="2021-02-21T06:00:00" tod="-1"><values t="-10" p="749" ws="1" wd="0" hum="79" hi="-10" cl="2" pt="0" pr="0" prflt="0" ts="0"
+                                icon="n.c3" descr="Cloudy"
+                                grade="4"  /></forecast><forecast valid="2021-02-21T09:00:00" tod="1"><values t="-10" p="748" ws="2" wd="5" hum="85" hi="-14" cl="3" pt="2" pr="1" prflt="0.1" ts="0"
+                                icon="c4.s1" descr="Mainly cloudy, light snow, brume"
+                                grade="4"  /></forecast><forecast valid="2021-02-21T12:00:00" tod="-1"><values t="-8" p="748" ws="2" wd="5" hum="79" hi="-12" cl="3" pt="2" pr="2" prflt="0.5" ts="0"
+                                icon="c4.s2" descr="Mainly cloudy, heavy snow"
+                                grade="4"  /></forecast><forecast valid="2021-02-21T15:00:00" tod="2"><values t="-7" p="746" ws="3" wd="5" hum="86" hi="-12" cl="3" pt="2" pr="3" prflt="2.2" ts="0"
+                                icon="c4.s3" descr="Mainly cloudy, very heavy snow"
+                                grade="3"  /></forecast><forecast valid="2021-02-21T18:00:00" tod="-1"><values t="-8" p="747" ws="2" wd="4" hum="88" hi="-12" cl="3" pt="2" pr="3" prflt="2.2" ts="0"
+                                icon="c4.s3" descr="Mainly cloudy, very heavy snow"
+                                grade="3"  gust_speed="5" /></forecast><forecast valid="2021-02-21T21:00:00" tod="3"><values t="-8" p="747" ws="2" wd="3" hum="89" hi="-13" cl="3" pt="2" pr="2" prflt="0.6" ts="0"
+                                icon="c4.s2" descr="Mainly cloudy, heavy snow"
+                                grade="2"  gust_speed="4" /></forecast></day><day date="2021-02-22" risem="456" setm="1069" durm="613" sunrise="1613979360" sunset="1614016140" tmin="-21" tmax="-11" pmin="749"
+                 pmax="760" wsmin="4" wsmax="5" hummin="68" hummax="89" cl="101" pt="2" pr="1" ts="0" icon="d.c2.s1" descr="Partly cloudy, light snow" p="755"
+                 ws="5" wd="2" hum="78" grademax="3" prflt="0.8"  gust_speed="8"><forecast valid="2021-02-22T00:00:00" tod="-1"><values t="-11" p="749" ws="4" wd="2" hum="89" hi="-17" cl="3" pt="2" pr="1" prflt="0.1" ts="0"
+                                icon="c4.s1" descr="Mainly cloudy, light snow"
+                                grade="3"  gust_speed="6" /></forecast><forecast valid="2021-02-22T03:00:00" tod="0"><values t="-11" p="750" ws="5" wd="2" hum="86" hi="-18" cl="3" pt="0" pr="0" prflt="0" ts="0"
+                                icon="c4" descr="Mainly cloudy"
+                                grade="3"  gust_speed="8" /></forecast><forecast valid="2021-02-22T06:00:00" tod="-1"><values t="-14" p="752" ws="5" wd="2" hum="83" hi="-22" cl="3" pt="0" pr="0" prflt="0" ts="0"
+                                icon="c4" descr="Mainly cloudy"
+                                grade="3"  gust_speed="8" /></forecast><forecast valid="2021-02-22T09:00:00" tod="1"><values t="-15" p="753" ws="5" wd="3" hum="79" hi="-23" cl="3" pt="2" pr="1" prflt="0.1" ts="0"
+                                icon="c4.s1" descr="Mainly cloudy, light snow"
+                                grade="2"  gust_speed="8" /></forecast><forecast valid="2021-02-22T12:00:00" tod="-1"><values t="-15" p="754" ws="5" wd="3" hum="75" hi="-24" cl="3" pt="2" pr="1" prflt="0.3" ts="0"
+                                icon="c4.s1" descr="Mainly cloudy, light snow"
+                                grade="2"  gust_speed="8" /></forecast><forecast valid="2021-02-22T15:00:00" tod="2"><values t="-15" p="755" ws="5" wd="2" hum="68" hi="-23" cl="3" pt="2" pr="1" prflt="0.3" ts="0"
+                                icon="c4.s1" descr="Mainly cloudy, light snow"
+                                grade="2"  gust_speed="8" /></forecast><forecast valid="2021-02-22T18:00:00" tod="-1"><values t="-17" p="757" ws="5" wd="2" hum="69" hi="-26" cl="0" pt="0" pr="0" prflt="0" ts="0"
+                                icon="n" descr="Fair"
+                                grade="2"  gust_speed="8" /></forecast><forecast valid="2021-02-22T21:00:00" tod="3"><values t="-21" p="760" ws="5" wd="2" hum="74" hi="-30" cl="0" pt="0" pr="0" prflt="0" ts="0"
+                                icon="n" descr="Fair"
+                                grade="2"  gust_speed="8" /></forecast></day><day date="2021-02-23" risem="454" setm="1071" durm="617" sunrise="1614065640" sunset="1614102660" tmin="-27" tmax="-23" pmin="762"
+                 pmax="767" wsmin="1" wsmax="4" hummin="60" hummax="74" cl="0" pt="0" pr="0" ts="0" icon="d" descr="Fair" p="0"
+                 ws="4" wd="2" hum="68" grademax="3" prflt="0"  gust_speed="7"></day><day date="2021-02-24" risem="451" setm="1073" durm="622" sunrise="1614151860" sunset="1614189180" tmin="-27" tmax="-15" pmin="758"
+                 pmax="767" wsmin="2" wsmax="5" hummin="59" hummax="73" cl="3" pt="2" pr="1" ts="0" icon="c4.s1" descr="Mainly cloudy, light snow" p="0"
+                 ws="5" wd="5" hum="66" grademax="2" prflt="0.1"  gust_speed="9"></day><day date="2021-02-25" risem="449" setm="1075" durm="626" sunrise="1614238140" sunset="1614275700" tmin="-14" tmax="2" pmin="745"
+                 pmax="755" wsmin="5" wsmax="6" hummin="78" hummax="91" cl="3" pt="2" pr="2" ts="0" icon="c4.rs2" descr="Mainly cloudy, heavy wet snow" p="0"
+                 ws="6" wd="6" hum="86" grademax="2" prflt="4.9"  gust_speed="10"></day><day date="2021-02-26" risem="447" setm="1077" durm="630" sunrise="1614324420" sunset="1614362220" tmin="2" tmax="4" pmin="740"
+                 pmax="745" wsmin="5" wsmax="7" hummin="81" hummax="95" cl="2" pt="1" pr="1" ts="0" icon="d.c3.r1" descr="Cloudy, light rain" p="0"
+                 ws="7" wd="7" hum="89" grademax="2" prflt="0.3"  gust_speed="12"></day><day date="2021-02-27" risem="444" setm="1079" durm="635" sunrise="1614410640" sunset="1614448740" tmin="0" tmax="2" pmin="740"
+                 pmax="740" wsmin="4" wsmax="6" hummin="82" hummax="90" cl="3" pt="0" pr="0" ts="0" icon="c4" descr="Mainly cloudy" p="0"
+                 ws="6" wd="7" hum="87" grademax="2" prflt="0"  gust_speed="11"></day></location></weather>
