+++ tests/test_init.py
@@ -1,0 +1,13 @@
+"""Tests for GisMeteo integration."""
+# from homeassistant.const import STATE_UNAVAILABLE
+#
+# from tests import init_integration
+#
+# async def test_async_setup_entry(hass):
+#     """Test a successful setup entry."""
+#     await init_integration(hass)
+#
+#     state = hass.states.get("weather.home")
+#     assert state is not None
+#     assert state.state != STATE_UNAVAILABLE
+#     assert state.state == "sunny"
